backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0
django-cors-headers>=3.14
openai>=0.27.0
python-pptx>=0.6.21
PyPDF2>=3.0.0
mammoth>=1.4.2
